% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{A function to do simple linear regression - the first of my first package - how cool!
This function will estimate the OLS estmates of beta, as well as the SD, t statistics, and hypothesis conclusions}
\usage{
linear(X, y, need_intercept = T)
}
\arguments{
\item{X}{n by p design matrix of numeric covariates where n is the sample size and p is the parameter space.}

\item{y}{n by 1 vector of numeric continuous outcomes}

\item{need_intercept}{do you need us to add an intercept column to your design matrix? we assume TRUE}
}
\description{
A function to do simple linear regression - the first of my first package - how cool!
This function will estimate the OLS estmates of beta, as well as the SD, t statistics, and hypothesis conclusions
}
\examples{
data(mtcars)
y = mtcars$mpg
X = mtcars$cyl
fit = linear(X,y)
betas = fit$coeffs

}
