% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{A function to do simple linear regression - the first of my first package - how cool!
This function will estimate the OLS estmates of beta, as well as the SD, t statistics, and hypothesis conclusions}
\usage{
linear(X, y, add_intercept = T, to_predict = NULL)
}
\arguments{
\item{X}{n by p design matrix of numeric covariates where n is the sample size and p is the parameter space.}

\item{y}{n by 1 vector of numeric continuous outcomes}

\item{add_intercept}{do you need to add an intercept column to your design matrix? we assume TRUE}

\item{to_predict}{an optional obervation(s) for which you would like the outcomes predicted}
}
\value{
coefficients, fitted values, residuals, predictions (if you asked for them),
t- statistics and conclusions to the hypothesis test that the coefficient is 0.
}
\description{
A function to do simple linear regression - the first of my first package - how cool!
This function will estimate the OLS estmates of beta, as well as the SD, t statistics, and hypothesis conclusions
}
\examples{

data(mtcars)
y = mtcars$mpg
X = mtcars$cyl
# can also make X a matrix, no problem
fit = linear(X,y) # this assumes your design matrix does not have an intercept in it
betas = fit$coeffs # these are your OLS regression coefficients

}
